+++ lib/util/connector.py
@@ -7,10 +7,7 @@ x
 from pymongo import MongoClient
 
 
-# Constants
-
-DEFAULT_NAME = "database"
-PROTOCOL = "mongodb"
+# Local Imports
 
 from lib.match import Match
 
@@ -20,34 +17,15 @@ class Connector( object ):
     x
     """
 
-    def __init__( self, address: str, port: int ):
-        host = f"{PROTOCOL}://{address}:{port}/"
-        self.client = MongoClient( host )
-        self.database = self.client[DEFAULT_NAME]
-        self.events = self.database['events']
-        self.history = self.database['history']
-        self.sessions = self.database['sessions']
-        self.matches = self.database['matches']
-        self.rounds = self.database['rounds']
-        self.teams = self.database['teams']
-        self.players = self.database['players']
-        self.stats = self.database['stats']
-        self.wstats = self.database['wstats']
+    def __init__( self, address: str, port: int = 27017 ):
+        """
+        Parameters
+        ----------
+        address : str
+            The address of a MongoDB database.
+        port : int
+            The port that the database runs on.
+        """
 
-
-    def get_all( self ) -> list:
-        sessions = []
-        pass
-        return sessions
-
-
-    def get_ranked( self ) -> list:
-        sessions = []
-        pass
-        return sessions
-
-
-    def get_unranked( self ) -> list:
-        sessions = []
-        pass
-        return sessions
+        self.host = f"mongodb://{address}:{port}/"
+        self.client = MongoClient( self.host )
