--- lib/util/helper.py
@@ -1,16 +1,0 @@
-"""
-x
-"""
-
-# Global Imports
-
-
-# Local Imports
-
-
-class Helper( object ):
-    """
-    x
-    """
-
-    pass
