+++ lib/util/parser.py
@@ -2,10 +2,16 @@
 x
 """
 
-# Global Imports
-
-
 # Local Imports
+
+from lib.event import Event
+from lib.match import Match
+from lib.player import Player
+from lib.round import Round
+from lib.session import Session
+from lib.table import Table
+from lib.team import Team
+from lib.util.connector import Connector
 
 
 class Parser( object ):
@@ -14,3 +20,32 @@ class Parser( object ):
     """
 
     pass
+
+    def get_all( self ) -> list:
+        """
+        x
+        """
+
+        sessions = []
+        pass
+        return sessions
+
+
+    def get_ranked( self ) -> list:
+        """
+        x
+        """
+
+        sessions = []
+        pass
+        return sessions
+
+
+    def get_unranked( self ) -> list:
+        """
+        x
+        """
+
+        sessions = []
+        pass
+        return sessions
