--- lib/util/validator.py
@@ -1,16 +1,0 @@
-"""
-x
-"""
-
-# Global Imports
-
-
-# Local Imports
-
-
-class Validator( object ):
-    """
-    x
-    """
-
-    pass
